+++ Projeto_de_Programa-o_N04_Grupo-01/backend/src/main/java/br/com/unit/classes/EventoAdministrado.java
@@ -2,7 +2,7 @@ package br.com.unit.classes;
 
 public class EventoAdministrado extends Evento{
 
-    protected EventoAdministrado(int idEvento, String nomeEvento, String descricaoEvento, String dataInicio, String dataFim, int capacidade) {
+    protected EventoAdministrado(long idEvento, String nomeEvento, String descricaoEvento, String dataInicio, String dataFim, int capacidade) {
         super(idEvento, nomeEvento, descricaoEvento, dataInicio, dataFim, capacidade);
     }
